requirements.txt
telebot
pillow
python-dotenv
pyTelegramBotAPI